Staging Information is only displayed if available
If staging is available and complete (defined), it is displayed as before. If not, a red box that reads 'Missing staging information' is displayed instead. I also corrected the spacing of major app components to take advantage of unused screen real estate. This created more room for the Data Summary panel.